chore(cleanup): delete dead field-day shell + orphan routes; extract shared types
Deletion-first close-out of the discoverability audit (reviewed via /autoplan;
plan + audit trail in the session plan file):

- New lib/my-work/field-day-types.ts holds OpenSession, VehicleOption, and
  ActivityEntryDto (previously defined inside page components and imported
  by lib — client/server boundary smell). Seven importers repointed.
- Delete WorkdayPanel.tsx (878 lines, never mounted — every import was
  type-only) and BusinessDayBar.tsx (only rendered inside it). Day lifecycle
  is unaffected: clock-in opens the business day, Day Review's CloseButton
  closes/reopens via /business-day/transition (kept).
- Delete caller-less routes: business-day/current (integration test helper
  repointed to a direct DB query first), operations/state (lib + unit tests
  stay), mileage + mileage/[id] (legacy wrappers; live /app/mileage pages
  use /api/v1/sessions/*).
- Delete MaterialHandlingPanel, DayReviewDatePicker, lib/active-job-ui.ts,
  lib/dates.ts (zero importers). Fix stale BusinessDayBar comments in ClockBar.

Gate: lint, typecheck, build, unit (1301) green. Tier-3 integration run
against an ephemeral seeded stack: 108 pass / 11 skip / 5 pre-existing
failures unrelated to this diff (stale tests; tracked separately).
