--- apps/web/lib/dates.ts
@@ -1,11 +1,0 @@
-/** Normalize DB timestamps (string or Date) to YYYY-MM-DD for queries and comparisons. */
-export function toDateOnly(value: string | Date | null | undefined): string {
-  if (!value) {
-    const now = new Date();
-    return `${now.getFullYear()}-${String(now.getMonth() + 1).padStart(2, "0")}-${String(now.getDate()).padStart(2, "0")}`;
-  }
-  if (value instanceof Date) {
-    return `${value.getFullYear()}-${String(value.getMonth() + 1).padStart(2, "0")}-${String(value.getDate()).padStart(2, "0")}`;
-  }
-  return String(value).slice(0, 10);
-}
